dao model classes brought upto date
