+++ server/src/dao/model/outgoing_classification.py
@@ -1,26 +1,63 @@
-
-
-class submitted_target:
+class outgoing_classification(object):
     """
-    This model class helps represent the final form of an AUVSI target submission.
-    Essentially just removes helper columns from outgoing_ classification models
+    Model class for outgoing classification table entries. This model describes
+    and stores the common column elements between the 'outgoing_manual' and 
+    'outgoing_autonomous' tables. At the time of writing, they are identical. In
+    the future if it is decided they need to diverge, describe these diverging
+    values in the corresponding child classes (outgoing_manual or outgoing_autonomous)
     """
 
-    # A list of keys to exclude in the output .json file generated by the DAO
-    # for standard and off_axis target types
-    KEYS_TO_EXCLUDE = ['id', 'target', 'crop_id', 'crop_path', 'submitted', 'description']
+    def __init__(self, tableValues=None, json=None):
+        """
+        Accepts various formats to instantiate this model object
 
-    # a list of keys to include in the output .json file generated by the DAO
-    # for emergent target type
-    EMERGENT_KEYS_TO_INCLUDE = ['latitude', 'longitude', 'description', 'autonomous']
+        @type tableValues: [object]
+        @param tableValues: List of table values, in table column order
 
-    def __init__(self, outgoingManualOrAutonomous, manual):
-        outDict = outgoingManualOrAutonomous.toDict()
+        @type json: {object}
+        @param json: Json dictionary of table values. Used by the REST API when receiving data
+        """
+        if tableValues is not None:
+            self.class_id = tableValues[0]
+            self.crop_id = tableValues[1]
+            self.target = tableValues[2]
+            self.type = tableValues[3]
+            self.latitude = tableValues[4]
+            self.longitude = tableValues[5]
+            self.orientation = tableValues[6]
+            self.shape = tableValues[7]
+            self.background_color = tableValues[8]
+            self.alphanumeric = tableValues[9]
+            self.alphanumeric_color = tableValues[10]
+            self.description = tableValues[11]
+            self.submitted = tableValues[12]
+        elif json is not None:
+            for prop in self.allProps():
+                if prop in json:
+                    setattr(self, prop, json[prop])
 
-        self.autonomous = not manual
-        for key in self.allProps():
-            if key in outDict:
-                setattr(self, key, outDict[key])
+    @property
+    def class_id(self):
+        """
+        Table id. Internal to the dao, not exposed by the REST API
+        """
+        return self._class_id
+
+    @class_id.setter
+    def class_id(self, id):
+        self._class_id = id
+
+    @property
+    def crop_id(self):
+        """
+        Unique crop_id, publicly exposed by the API and used to access information on the 
+        image in various states (raw, cropped, and classified)
+        """
+        return self._crop_id
+
+    @crop_id.setter
+    def crop_id(self, crop_id):
+        self._crop_id = crop_id
 
     @property
     def target(self):
@@ -36,20 +73,6 @@ class submitted_target:
     @target.setter
     def target(self, target):
         self._target = target
-
-    @property
-    def crop_id(self):
-        """
-        crop_id references the id the of the cropped image to use for submission.
-        Information on this cropped image is held in either the cropped_autonomous
-        or cropped_manual table (depending on if this is a manual or autonomous
-        submission)
-        """
-        return self._crop_id
-
-    @crop_id.setter
-    def crop_id(self, crop_id):
-        self._crop_id = crop_id
 
     @property
     def type(self):
@@ -162,31 +185,43 @@ class submitted_target:
         self._description = description
 
     @property
-    def autonomous(self):
-        return self._autonomous
+    def submitted(self):
+        """
+        Boolean to indicate whether the classification has been submitted to the judges yet
+        """
+        return self._submitted
 
-    @autonomous.setter
-    def autonomous(self, autonomous):
-        self._autonomous = autonomous
+    @submitted.setter
+    def submitted(self, submitted):
+        self._submitted = submitted
 
     # TODO: this is hacky and i hate it
     def allProps(self):
-        return ['target', 'crop_id', 'type', 'latitude', 'longitude', 'orientation', 'shape', 'background_color', 'alphanumeric', 'alphanumeric_color', 'description', 'autonomous']
+        return ['class_id', 'crop_id', 'type', 'target', 'latitude', 'longitude', 'orientation', 'shape', 'background_color', 'alphanumeric', 'alphanumeric_color', 'description', 'submitted']
 
-    def toJson(self):
+    def toDict(self, exclude=None, include=None):
+        """
+        Return attributes contained in this model as a dictionary
+
+        @type exclude: (string)
+        @param exclude: Attribute names to exclude from the generated result. Exclude takes priority over include
+
+        @param include: (string)
+        @param include: as opposed to exclude, this allows you to specify only specific keys to include in the generated result
+
+        @rtype: {string}
+        @return: String dictionary of classification properties
+
+        """
         dict = {}
-        if self.type == 'emergent':
-            # generate an emergent target output
-            for attr, value in self.__dict__.items():
-                corrected_name = attr[1:] # remove first underscore
+        for attr, value in self.__dict__.items():
+            corrected_name = attr[1:] # remove first underscore
+            
+            if exclude is not None and corrected_name in exclude:
+                continue
 
-                if corrected_name in self.EMERGENT_KEYS_TO_INCLUDE:
-                    dict[corrected_name] = value
-        else:
-            for attr, value in self.__dict__.items():
-                corrected_name = attr[1:] # remove first underscore
-
-                if corrected_name not in self.KEYS_TO_EXCLUDE:
-                    dict[corrected_name] = value            
-    
+            if include is not None and corrected_name in include:
+                dict[corrected_name] = value
+            elif include is None:
+                dict[corrected_name] = value
         return dict
